% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcsis-package.R
\docType{package}
\name{cdcsis-package}
\alias{cdcsis-package}
\title{Conditional Distance Correlation and Its Related Feature Screening Method}
\description{
Gives conditional distance correlation and performs the conditional distance correlation sure independence screening procedure for ultrahigh dimensional data. 
The conditional distance correlation is a novel conditional dependence measurement of two random variables given a third variable. 
The conditional distance correlation sure independence screening is used for screening variables in ultrahigh dimensional setting.
}
\examples{
set.seed(0)
n <- 100 # sample size
p <- 10 # dimensionality
rho <- 0.5 # the correlation between pairwise predictors.
Sigma <- matrix(rho, p, p)
diag(Sigma) <- 1

require(MASS)
x <- mvrnorm(n,rep(0,p), Sigma)

y <- x[,2] + x[,3] + rnorm(n)
z <- x[,1]
cdcsis(x,y,z,2)
}
\references{
Canhong Wen, Wenliang Pan, Mian Huang and Xueqin Wang(2014). Conditional distance correlation sure independence screening for ultrahigh dimensional data.
}
\author{
Canhong Wen, Wenliang Pan, Mian Huang, and Xueqin Wang

Maintainer: Canhong Wen <wencanhong@gmail.com>
}
\keyword{SIS,}
\keyword{confounder}
\keyword{correlation,}
\keyword{distance}
