% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcsis.R
\name{cdcsis}
\alias{cdcsis}
\title{Conditional Distance Correlation Sure Independence Screening (CDCSIS)}
\usage{
cdcsis(x, y, z, thres)
}
\arguments{
\item{x}{a matrix}

\item{y}{a numeric vector or matrix with compatible dimensions to \code{x}}

\item{z}{the variable being conditioned. \code{z} is a numeric vector or
matrix with compatible dimensions to \code{x}}

\item{thres}{the threshold of the number of pedictors recuited by CDCSIS.
Should be less than or equal than the number of column of \code{x}.}
}
\value{
\item{ CDCSISind }{the vector of indices selected by CDCSIS } \item{
thres }{the threshold of the number of pedictors recuited by CDCSIS} \item{
DC }{ the distance correlation for each dimensionality of \code{x}} \item{
DCord }{the order of \code{DC} for each dimensionality of \code{x} }
}
\description{
Performs conditional distance correlation sure independence screening
(CDCSIS).
}
\details{
It performs conditional distance correlation sure independence screening
(CDCSIS).
}
\examples{
set.seed(0)
n <- 100
p <- 10
rho <- 0.5
Sigma <- matrix(rho, p, p)
diag(Sigma) <- 1

require(MASS)
x <- mvrnorm(n,rep(0,p), Sigma)
y <- x[,2] + x[,3] + rnorm(n)
z <- x[,1]

cdcsis(x,y,z,2)

}
\references{
Canhong Wen, Wenliang Pan, Mian Huang and Xueqin Wang(2014).
Conditional distance correlation sure independence screening for ultrahigh
dimensional data. \emph{Submitted to Biostatistics}.
}
\seealso{
\code{\link{cdcor.ada}}, \code{\link{cdcor}}
}
\author{
Canhong Wen, Wenliang Pan, and Xueqin Wang
}
\keyword{Ultrahigh}
\keyword{conditional}
\keyword{correlation}
\keyword{dimensional}
\keyword{distance}
\keyword{independence}
\keyword{screening}
\keyword{sure}
