% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcor.R
\name{cdcor}
\alias{cdcor}
\title{Conditional Distance Correlation with Given Bandwidth}
\usage{
cdcor(x, y, z, width, index = 1)
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{y}{a numeric vector or matrix with compatible dimensions to \code{x}}

\item{z}{the variable being conditioned. \code{z} is a numeric vector or
matrix with compatible dimensions to \code{x}}

\item{width}{a positive value, user-specified}

\item{index}{exponent on Euclidean distance, in (0,2]}
}
\value{
\item{ cdcov }{the conditional distance covariance with given
variable z } \item{ mcdcov }{mean of the conditional distance covariance
with given variable z} \item{ width }{ the bandwidth } \item{ index }{
exponent on Euclidean distance, in (0,2] }
}
\description{
Performs conditional distance correlation with given bandwidth.
}
\details{
It performs conditional distance correlation with given bandwidth.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
z <- rnorm(100)
cdcor(x,y,z,0.25)

}
\references{
Canhong Wen, Wenliang Pan, Mian Huang and Xueqin Wang(2014).
Conditional distance correlation sure independence screening for ultrahigh
dimensional data. \emph{Submitted to Biostatistics}.
}
\seealso{
\code{\link{cdcor.ada}}, \code{\link{cdcov}}
}
\author{
Canhong Wen, Wenliang Pan, Mian Huang, and Xueqin Wang
}
\keyword{conditional}
\keyword{correlation}
\keyword{distance}
