% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcov.R
\name{cdcov}
\alias{cdcov}
\title{Conditional Distance Covariance with Given Bandwidth}
\usage{
cdcov(x, y, z, width, index = 1)
}
\arguments{
\item{x}{a numeric vector or matrix}

\item{y}{a numeric vector or matrix with compatible dimensions to \code{x}}

\item{z}{the variable being conditioned. \code{z} is a numeric vector or
matrix with compatible dimensions to \code{x}}

\item{width}{a positive value, user-specified}

\item{index}{exponent on Euclidean distance, in (0,2]}
}
\value{
\item{cdcov}{conditional distance covariance with given variable
\code{z}; has the same length with \code{z}} \item{width}{the bandwidth}
\item{index}{ exponent on Euclidean distance, in (0,2] }
}
\description{
Performs conditional distance covariance with given bandwidth.
}
\details{
Perform the measurement of the conditional independence between x and y
given z with given bandwidth.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
z <- rnorm(100)
cdcov(x,y,z,0.25)
}
\references{
Canhong Wen, Wenliang Pan, Mian Huang and Xueqin Wang(2014).
Conditional distance correlation sure independence screening for ultrahigh
dimensional data. \emph{Submitted to Biostatistics}.
}
\seealso{
\code{\link{cdcor}}
}
\author{
Canhong Wen, Wenliang Pan, and Xueqin Wang
}
\keyword{conditional}
\keyword{covariance}
\keyword{distance}
