% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdcsis-package.R
\docType{package}
\name{cdcsis-package}
\alias{cdcsis-package}
\title{Conditional Distance Correlation Based Feature Screening and Conditional Independence Inference}
\description{
Conditional distance correlation <doi:10.1080/01621459.2014.993081> is a novel conditional dependence measurement of two multivariate random variables given a confounding variable. 
This package provides conditional distance correlation, performs the conditional distance correlation sure independence screening procedure for ultrahigh dimensional data <doi:10.5705/ss.202014.0117>, 
and conducts conditional distance covariance test for conditional independence assumption of two multivariate variable.
}
\references{
Wang, X., Pan, W., Hu, W., Tian, Y. and Zhang, H., 2015. Conditional distance correlation. Journal of the American Statistical Association, 110(512), pp.1726-1734.

Wen, C., Pan, W., Huang, M. and Wang, X., 2018. Sure independence screening adjusted for confounding covariates with ultrahigh-dimensional data. Statistica Sinica, 28, pp.293-317.
}
\author{
Wenhao Hu, Mian Huang, Wenliang Pan, Xueqin Wang, Canhong Wen, Yuan Tian, Heping Zhang, Jin Zhu
Maintainer: Jin Zhu <zhuj37@mail2.sysu.edu.cn>
}
